Improve asr error handling.
When an upstream nameserver is not available asr is not synthesizing a
SERVFAIL rcode (duh), but sets ar_errno. When we need SERVFAIL further
down,we need to set the rcode ourselves.

While here, don't complain about a too short packet when asr already
told us that resolving did not work out in check_dns64_done.

OK semarie