Remove some (soon to be) outdated documentation
The documentation of the BN_MOD_CTX has been out of sync with reality
for decades. The structure is now opaque, so its members should not be
documented this way. They internals aren't important for the rest of
the page.

BN_MOD_CTX_init() will soon be removed. It's useless unless you like
leaks.