clockintr: remove support for independent schedclock()
Remove the scaffolding for an independent schedclock().  With the
removal of the independent schedclock() from alpha, schedhz is zero on
all platforms and this schedclock-specific code is now unused.

It is possible that schedclock() will repurposed for use in the
future.  Even if this happens, the schedclock handle will not live in
struct clockintr_queue.