clockintr: rename "clockintr_cancel_locked" to "clockqueue_pend_delete"
The function clockintr_cancel_locked() is poorly named.  "Cancel" is a
broad word that connotes too much.  "clockintr" implies that we are
mainly operating on the clockintr itself, which is untrue: cq_pend,
the particular data structure, is important.

clockqueue_pend_delete() is a much better name: it specifies the data
structure we're operating on ("cq_pend") and the fundamental operation
at work ("delete").