Remove a bunch of function pointers from X509_STORE_CTX
These are only ever set to one particular function which is either local
to this file or part of the public API and we never added the public API
to set them to something else. Prefix the local functions touched in this
commit with x509_vfy_. More cleanup to follow.

ok joshua jsing